Factor out some more code
